add lnurl.ts & lint
