+++ src/screens/QRScan/index.tsx
@@ -13,7 +13,8 @@ import { useThemeContext } from '@src/context/Theme'
 import { NS } from '@src/i18n'
 import { getCustomMintNames, getDefaultMint } from '@store/mintStore'
 import { globals, mainColors } from '@styles'
-import { decodeLnInvoice, extractStrFromURL, hasTrustedMint, isCashuToken, isLnurlOrAddress, isNull, isStr, isUrl } from '@util'
+import { decodeLnInvoice, extractStrFromURL, hasTrustedMint, isCashuToken, isNull, isStr, isUrl } from '@util'
+import { isLnurlOrAddress } from '@util/lnurl'
 import { getTokenInfo } from '@wallet/proofs'
 import { BarCodeScanner, PermissionStatus } from 'expo-barcode-scanner'
 import { Camera, FlashMode } from 'expo-camera'
@@ -114,27 +115,28 @@ export default function QRScanPage({ navigation, route }: TQRScanPageProps) {
 			return navigation.navigate('mint confirm', { mintUrl: data })
 		}
 		// handle LNURL
-
-		if (isLnurlOrAddress(data) ) {
-	
-			if (mint === undefined || balance === undefined) {
-
+		if (isLnurlOrAddress(data)) {
+			if (!mint || !balance) {
 				// user has not selected the mint yet (Pressed scan QR and scanned a Lightning invoice)
 				const mintsWithBal = await getMintsBalances()
 				const mints = await getCustomMintNames(mintsWithBal.map(m => ({ mintUrl: m.mintUrl })))
 				const nonEmptyMint = mintsWithBal.filter(m => m.amount > 0)
+				// user has no funds
+				if (!nonEmptyMint.length) {
+					// user is redirected to the mint selection screen where he gets an appropriate message
+					return navigation.navigate('selectMint', {
+						mints,
+						mintsWithBal,
+						isMelt: true,
+						allMintsEmpty: true,
+						scanned: true
+					})
+				}
 				const mintUsing = mints.find(m => m.mintUrl === nonEmptyMint[0].mintUrl) || { mintUrl: 'N/A', customName: 'N/A' }
-				
-				return navigation.navigate('selectAmount', { mint:mintUsing, balance:nonEmptyMint[0].amount, isMelt: true, lnurl: data })		
-
-
-			} 
-			
-			return navigation.navigate('selectAmount', { mint, balance, isMelt: true, lnurl: data })	
-			
-			
+				return navigation.navigate('selectAmount', { mint: mintUsing, balance: nonEmptyMint[0].amount, isMelt: true, lnurl: data })
+			}
+			return navigation.navigate('selectAmount', { mint, balance, isMelt: true, lnurl: data })
 		}
-
 		// handle LN invoice
 		try {
 			const invoice = extractStrFromURL(data) || data
